Effectful.Log.Static

Running with logging

withLogFunc :: MonadUnliftIO m => LogOptions -> (LogFunc -> m a) -> m a

Given a LogOptions value, run the given function with the specified LogFunc. A common way to use this function is:

let isVerbose = False -- get from the command line instead
logOptions' <- logOptionsHandle stderr isVerbose
let logOptions = setLogUseTime True logOptions'
withLogFunc logOptions $ \lf -> do
  let app = App -- application specific environment
        { appLogFunc = lf
        , appOtherStuff = ...
        }
  runRIO app $ do
    logInfo "Starting app"
    myApp

Since: rio-0.0.0.0

newLogFunc :: (MonadIO n, MonadIO m) => LogOptions -> n (LogFunc, m ())

Given a LogOptions value, returns both a new LogFunc and a sub-routine that disposes it.

Intended for use if you want to deal with the teardown of LogFunc yourself, otherwise prefer the withLogFunc function instead.

Since: rio-0.1.3.0

data LogFunc

A logging function, wrapped in a newtype for better error messages.

An implementation may choose any behavior of this value it wishes, including printing to standard output or no action at all.

Since: rio-0.0.0.0

data Log :: Effect

Provide the ability to log messages via LogFunc like RIO.

data family StaticRep (e :: Effect)

Internal representations of statically dispatched effects.

logOptionsHandle

Arguments

:: MonadIO m

=> Handle

-> Bool
Verbose Flag

-> m LogOptions

Create a LogOptions value from the given Handle and whether to perform verbose logging or not. Individiual settings can be overridden using appropriate set functions. Logging output is guaranteed to be non-interleaved only for a UTF-8 Handle in a multi-thread environment.

When Verbose Flag is True, the following happens:

setLogVerboseFormat is called with True

setLogUseColor is called with True (except on Windows)

setLogUseLoc is called with True

setLogUseTime is called with True

setLogMinLevel is called with Debug log level

Since: rio-0.0.0.0

runLogRaw :: IOE :> es => LogFunc -> Eff (Log : es) a -> Eff es a

Run the Log effect.

runLog :: forall es a. IOE :> es => LogOptions -> Eff (Log : es) a -> Eff es a

Run the Log effect with withLogFunc.

Log options

data LogOptions

Configuration for how to create a LogFunc. Intended to be used with the withLogFunc function.

Since: rio-0.0.0.0

setLogMinLevel :: LogLevel -> LogOptions -> LogOptions

Set the minimum log level. Messages below this level will not be printed.

Default: in verbose mode, LevelDebug. Otherwise, LevelInfo.

Since: rio-0.0.0.0

setLogMinLevelIO :: IO LogLevel -> LogOptions -> LogOptions

Refer to setLogMinLevel. This modifier allows to alter the verbose format value dynamically at runtime.

Default: in verbose mode, LevelDebug. Otherwise, LevelInfo.

Since: rio-0.1.3.0

setLogVerboseFormat :: Bool -> LogOptions -> LogOptions

Use the verbose format for printing log messages.

Default: follows the value of the verbose flag.

Since: rio-0.0.0.0

setLogVerboseFormatIO :: IO Bool -> LogOptions -> LogOptions

Refer to setLogVerboseFormat. This modifier allows to alter the verbose format value dynamically at runtime.

Default: follows the value of the verbose flag.

Since: rio-0.1.3.0

setLogTerminal :: Bool -> LogOptions -> LogOptions

Do we treat output as a terminal. If True, we will enable sticky logging functionality.

Default: checks if the Handle provided to logOptionsHandle is a terminal with hIsTerminalDevice.

Since: rio-0.0.0.0

setLogUseTime :: Bool -> LogOptions -> LogOptions

Include the time when printing log messages.

Default: True in debug mode, False otherwise.

Since: rio-0.0.0.0

setLogUseColor :: Bool -> LogOptions -> LogOptions

Use ANSI color codes in the log output.

Default: True if in verbose mode and the Handle is a terminal device.

Since: rio-0.0.0.0

setLogUseLoc :: Bool -> LogOptions -> LogOptions

Use code location in the log output.

Default: True if in verbose mode, False otherwise.

Since: rio-0.1.2.0

setLogFormat :: (Utf8Builder -> Utf8Builder) -> LogOptions -> LogOptions

Set format method for messages

Default: id

Since: rio-0.1.13.0

setLogLevelColors :: (LogLevel -> Utf8Builder) -> LogOptions -> LogOptions

ANSI color codes for LogLevel in the log output.

Default: LevelDebug = "\ESC[32m" -- Green LevelInfo = "\ESC[34m" -- Blue LevelWarn = "\ESC[33m" -- Yellow LevelError = "\ESC[31m" -- Red LevelOther _ = "\ESC[35m" -- Magenta

Since: rio-0.1.18.0

setLogSecondaryColor :: Utf8Builder -> LogOptions -> LogOptions

ANSI color codes for secondary content in the log output.

Default: "\ESC[90m" -- Bright black (gray)

Since: rio-0.1.18.0

setLogAccentColors

Arguments

:: (Int -> Utf8Builder)
This should be a total function.

-> LogOptions

-> LogOptions

ANSI color codes for accents in the log output. Accent colors are indexed by Int.

Default: const "\ESC[92m" -- Bright green, for all indicies

Since: rio-0.1.18.0

Standard logging functions

logDebug :: (Log :> es, HasCallStack) => Utf8Builder -> Eff es ()

Lifted logDebug.

logInfo :: (Log :> es, HasCallStack) => Utf8Builder -> Eff es ()

Lifted logInfo.

logWarn :: (Log :> es, HasCallStack) => Utf8Builder -> Eff es ()

Lifted logWarn.

logError :: (Log :> es, HasCallStack) => Utf8Builder -> Eff es ()

Lifted logError.

logOther

Arguments

:: (Log :> es, HasCallStack)

=> Text
level

-> Utf8Builder

-> Eff es ()

Lifted logOther.

Advanced logging functions

Sticky logging

logSticky :: (Log :> es, HasCallStack) => Utf8Builder -> Eff es ()

Lifted logSticky.

logStickyDone :: (Log :> es, HasCallStack) => Utf8Builder -> Eff es ()

Lifted logStickyDone.

With source

logDebugS :: (Log :> es, HasCallStack) => LogSource -> Utf8Builder -> Eff es ()

Lifted logDebugS.

logInfoS :: (Log :> es, HasCallStack) => LogSource -> Utf8Builder -> Eff es ()

Lifted logInfoS.

logWarnS :: (Log :> es, HasCallStack) => LogSource -> Utf8Builder -> Eff es ()

Lifted logWarnS.

logErrorS :: (Log :> es, HasCallStack) => LogSource -> Utf8Builder -> Eff es ()

Lifted logErrorS.

logOtherS

Arguments

:: (Log :> es, HasCallStack)

=> Text
level

-> LogSource

-> Utf8Builder

-> Eff es ()

Lifted logOtherS.

Generic log function

logGeneric :: (Log :> es, HasCallStack) => LogSource -> LogLevel -> Utf8Builder -> Eff es ()

Lifted logGeneric.

Advanced running functions

mkLogFunc :: (CallStack -> LogSource -> LogLevel -> Utf8Builder -> IO ()) -> LogFunc

Create a LogFunc from the given function.

Since: rio-0.0.0.0

logOptionsMemory :: MonadIO m => m (IORef Builder, LogOptions)

Create a LogOptions value which will store its data in memory. This is primarily intended for testing purposes. This will return both a LogOptions value and an IORef containing the resulting Builder value.

This will default to non-verbose settings and assume there is a terminal attached. These assumptions can be overridden using the appropriate set functions.

Since: rio-0.0.0.0

Data types

data LogLevel

The log level of a message.

Since: rio-0.0.0.0

Constructors

LevelDebug

LevelInfo

LevelWarn

LevelError

LevelOther !Text

type LogSource = Text

Where in the application a log message came from. Used for display purposes only.

Since: rio-0.0.0.0

data CallStack

CallStacks are a lightweight method of obtaining a partial call-stack at any point in the program.

A function can request its call-site with the HasCallStack constraint. For example, we can define

putStrLnWithCallStack :: HasCallStack => String -> IO ()

as a variant of putStrLn that will get its call-site and print it, along with the string given as argument. We can access the call-stack inside putStrLnWithCallStack with callStack.

>>> :{
putStrLnWithCallStack :: HasCallStack => String -> IO ()
putStrLnWithCallStack msg = do
  putStrLn msg
  putStrLn (prettyCallStack callStack)
:}

Thus, if we call putStrLnWithCallStack we will get a formatted call-stack alongside our string.

>>> putStrLnWithCallStack "hello"
hello
CallStack (from HasCallStack):
  putStrLnWithCallStack, called at <interactive>:... in interactive:Ghci...

GHC solves HasCallStack constraints in three steps:

If there is a CallStack in scope -- i.e. the enclosing function has a HasCallStack constraint -- GHC will append the new call-site to the existing CallStack.

If there is no CallStack in scope -- e.g. in the GHCi session above -- and the enclosing definition does not have an explicit type signature, GHC will infer a HasCallStack constraint for the enclosing definition (subject to the monomorphism restriction).

If there is no CallStack in scope and the enclosing definition has an explicit type signature, GHC will solve the HasCallStack constraint for the singleton CallStack containing just the current call-site.

CallStacks do not interact with the RTS and do not require compilation with -prof. On the other hand, as they are built up explicitly via the HasCallStack constraints, they will generally not contain as much information as the simulated call-stacks maintained by the RTS.

A CallStack is a [(String, SrcLoc)]. The String is the name of function that was called, the SrcLoc is the call-site. The list is ordered with the most recently called function at the head.

NOTE: The intrepid user may notice that HasCallStack is just an alias for an implicit parameter ?callStack :: CallStack. This is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.8.1.0

Convenience functions

displayCallStack :: CallStack -> Utf8Builder

Convert a CallStack value into a Utf8Builder indicating the first source location.

TODO Consider showing the entire call stack instead.

Since: rio-0.0.0.0

Type-generic logger

data GLog msg :: Effect

Provide the ability to log messages via GLogFunc like RIO.

runGLogRaw :: IOE :> es => GLogFunc msg -> Eff (GLog msg : es) a -> Eff es a

Run the GLog effect.

glog :: GLog msg :> es => msg -> Eff es ()

Lifted glog.

data GLogFunc msg

A generic logger of some type msg.

Your GLocFunc can re-use the existing classical logging framework of RIO, and/or implement additional transforms, filters. Alternatively, you may log to a JSON source in a database, or anywhere else as needed. You can decide how to log levels or severities based on the constructors in your type. You will normally determine this in your main app entry point.

Since: rio-0.1.13.0

gLogFuncClassic :: (HasLogLevel msg, HasLogSource msg, Display msg) => LogFunc -> GLogFunc msg

Make a GLogFunc via classic LogFunc. Use this if you'd like to log your generic data type via the classic RIO terminal logger.

Since: rio-0.1.13.0

mkGLogFunc :: (CallStack -> msg -> IO ()) -> GLogFunc msg

Make a custom generic logger. With this you could, for example, write to a database or a log digestion service. For example:

mkGLogFunc (\stack msg -> send (Data.Aeson.encode (JsonLog stack msg)))

Since: rio-0.1.13.0

contramapMaybeGLogFunc :: (a -> Maybe b) -> GLogFunc b -> GLogFunc a

A vesion of contramapMaybeGLogFunc which supports filering.

Since: rio-0.1.13.0

contramapGLogFunc :: (a -> b) -> GLogFunc b -> GLogFunc a

A contramap. Use this to wrap sub-loggers via mapRIO.

If you are on base > 4.12.0, you can just use contramap.

Since: rio-0.1.13.0

class HasLogLevel msg where

Level, if any, of your logs. If unknown, use LogOther. Use for your generic log data types that want to sit inside the classic log framework.

Since: rio-0.1.13.0

Methods

getLogLevel :: msg -> LogLevel

class HasLogSource msg where

Source of a log. This can be whatever you want. Use for your generic log data types that want to sit inside the classic log framework.

Since: rio-0.1.13.0

Methods

getLogSource :: msg -> LogSource